backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
qrcode==7.4.2
pillow==10.1.0
aiosmtplib==2.0.2
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3
PyJWT==2.8.0 